Added the Sliding option  and the color comparison to the pressure plates
I created a new script to get the color of the heavy object.
I created the prefab of the heavy object.
I added the option to slide the object once the pressure plate is pressed.
I added the color comparison between the pressure plate and the heavy object.
I added a box collider to the pressure plate prefab